# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: İ-EP.APP API
  description: |
    İ-EP.APP çoklu kiracılı okul yönetim sistemi için kapsamlı RESTful API.
    
    ## Özellikler
    - Çoklu kiracı mimarisi (Multi-tenant)
    - Role tabanlı erişim kontrolü (RBAC)
    - Güvenli dosya yönetimi
    - Türk eğitim sistemi desteği
    - Gerçek zamanlı bildirimler
    
    ## Kimlik Doğrulama
    API, NextAuth.js tabanlı kimlik doğrulama kullanır. Her istekte aşağıdaki header'lar gereklidir:
    - `X-User-Email`: Kullanıcı e-posta adresi
    - `X-User-ID`: Kullanıcı benzersiz ID'si
    - `x-tenant-id`: Kiracı (okul) ID'si
    
    ## Roller
    - `super_admin`: Sistem yöneticisi
    - `admin`: Okul yöneticisi
    - `teacher`: Öğretmen
    - `student`: Öğrenci
    - `parent`: Veli
  version: 1.0.0
  contact:
    name: İ-EP.APP API Desteği
    url: https://i-ep.app/support
    email: api-support@i-ep.app
  license:
    name: Proprietary
servers:
  - url: https://api.i-ep.app
    description: Production sunucusu
  - url: http://localhost:3000/api
    description: Development sunucusu
security:
  - ApiKeyAuth: []
  - TenantAuth: []
paths:
...
  /attendance:
    get:
      tags:
        - Attendance (Devamsızlık)
      summary: Devam kayıtlarını listele
      description: |
        Devam kayıtlarını filtreli olarak getirir.
        Öğretmenler kendi sınıflarını, öğrenciler kendi kayıtlarını görebilir.
      parameters:
        - name: studentId
          in: query
          description: Öğrenci ID'si ile filtrele
          schema:
            type: string
            format: uuid
        - name: classId
          in: query
          description: Sınıf ID'si ile filtrele
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          description: Belirli tarih (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          description: Başlangıç tarihi (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Bitiş tarihi (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: status
          in: query
          description: Devam durumu
          schema:
            $ref: '#/components/schemas/AttendanceStatus'
        - name: limit
          in: query
          description: Kayıt limiti
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          description: Başlangıç pozisyonu
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Devam kayıtları başarıyla getirildi
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttendanceRecord'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Attendance (Devamsızlık)
      summary: Devam kaydı oluştur
      description: |
        Tek veya toplu devam kaydı oluşturur.
        Toplu kayıt için 'attendance' dizisi kullanılır.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CreateAttendanceRequest'
                - $ref: '#/components/schemas/BulkAttendanceRequest'
      responses:
        '201':
          description: Devam kaydı başarıyla oluşturuldu
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/AttendanceRecord'
                      - type: array
                        items:
                          $ref: '#/components/schemas/AttendanceRecord'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Devam kaydı zaten mevcut
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # ========================================
  # GRADES (NOTLAR)
  # ========================================
...
  /classes:
    get:
      tags:
        - Classes (Sınıflar)
      summary: Sınıf listesini getir
      description: Okuldaki tüm sınıfları listeler
      parameters:
        - name: grade
          in: query
          description: Sınıf seviyesi ile filtrele
          schema:
            type: string
        - name: academic_year
          in: query
          description: Akademik yıl ile filtrele
          schema:
            type: string
            pattern: '^\d{4}-\d{4}$'
        - name: status
          in: query
          description: Sınıf durumu
          schema:
            $ref: '#/components/schemas/ClassStatus'
        - name: teacher_id
          in: query
          description: Sınıf öğretmeni ID'si
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sınıf listesi başarıyla getirildi
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Class'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Classes (Sınıflar)
      summary: Yeni sınıf oluştur
      description: Yeni bir sınıf oluşturur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassRequest'
      responses:
        '201':
          description: Sınıf başarıyla oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ========================================
  # STORAGE (DOSYA YÖNETİMİ)
  # ========================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Kullanıcı ID'si
    TenantAuth:
      type: apiKey
      in: header
      name: x-tenant-id
      description: Kiracı (okul) ID'si
  schemas:
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        student_name:
          type: string
          description: Öğrenci adı
        class_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/AttendanceStatus'
        time_in:
          type: string
          format: time
          description: Giriş saati
        time_out:
          type: string
          format: time
          description: Çıkış saati
        notes:
          type: string
          description: Notlar
        excuse_reason:
          type: string
          description: Mazeret nedeni
        excuse_document:
          type: string
          description: Mazeret belgesi
        marked_by:
          type: string
          format: uuid
          description: Kaydı yapan kişi ID'si
        tenant_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      required:
        - id
        - student_id
        - class_id
        - date
        - status
        - tenant_id
    AttendanceStatus:
      type: string
      enum:
        - present
        - absent
        - late
        - excused
        - sick
      description: |
        Devam durumları:
        - present: Mevcut
        - absent: Devamsız
        - late: Geç
        - excused: Mazeretli
        - sick: Hasta
    CreateAttendanceRequest:
      type: object
      properties:
        studentId:
          type: string
          format: uuid
        classId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/AttendanceStatus'
        timeIn:
          type: string
          format: time
        timeOut:
          type: string
          format: time
        notes:
          type: string
        excuseReason:
          type: string
        excuseDocument:
          type: string
      required:
        - studentId
        - classId
        - date
        - status
    BulkAttendanceRequest:
      type: object
      properties:
        classId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        attendance:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              studentId:
                type: string
                format: uuid
              status:
                $ref: '#/components/schemas/AttendanceStatus'
              timeIn:
                type: string
                format: time
              timeOut:
                type: string
                format: time
              notes:
                type: string
              excuseReason:
                type: string
            required:
              - studentId
              - status
      required:
        - classId
        - date
        - attendance

    # ========================================
    # GRADE SCHEMAS
    # ========================================
...
    Class:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Sınıf adı (örn. 5/A)
        grade:
          type: string
          description: Sınıf seviyesi
        section:
          type: string
          description: Şube
        capacity:
          type: integer
          description: Kapasite
          minimum: 1
          maximum: 50
        current_enrollment:
          type: integer
          description: Mevcut öğrenci sayısı
        academic_year:
          type: string
          pattern: '^\d{4}-\d{4}$'
        teacher_id:
          type: string
          format: uuid
          description: Sınıf öğretmeni ID'si
        room_number:
          type: string
          description: Sınıf numarası
        status:
          $ref: '#/components/schemas/ClassStatus'
        description:
          type: string
        tenant_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - grade
        - section
        - capacity
        - academic_year
        - status
        - tenant_id
    ClassStatus:
      type: string
      enum:
        - active
        - inactive
        - archived
      description: |
        Sınıf durumları:
        - active: Aktif
        - inactive: Pasif
        - archived: Arşivlendi
    CreateClassRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        grade:
          type: string
          minLength: 1
          maxLength: 10
        section:
          type: string
          minLength: 1
          maxLength: 10
        capacity:
          type: integer
          minimum: 1
          maximum: 50
        academic_year:
          type: string
          pattern: '^\d{4}-\d{4}$'
        teacher_id:
          type: string
          format: uuid
        room_number:
          type: string
        current_enrollment:
          type: integer
          minimum: 0
          default: 0
        status:
          $ref: '#/components/schemas/ClassStatus'
          default: active
        description:
          type: string
      required:
        - name
        - grade
        - section
        - capacity
        - academic_year

    # ========================================
    # STORAGE SCHEMAS
    # ========================================
...
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
          description: Toplam kayıt sayısı
        page:
          type: integer
          description: Mevcut sayfa
        limit:
          type: integer
          description: Sayfa başına kayıt
        totalPages:
          type: integer
          description: Toplam sayfa sayısı
        hasNextPage:
          type: boolean
          description: Sonraki sayfa var mı?
        hasPreviousPage:
          type: boolean
          description: Önceki sayfa var mı?
        offset:
          type: integer
          description: Başlangıç pozisyonu
        hasMore:
          type: boolean
          description: Daha fazla kayıt var mı?
      required:
        - total
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Hata mesajı
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Hatalı alan
              message:
                type: string
                description: Alan için hata mesajı
              code:
                type: string
                description: Hata kodu
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: İstek yolu
        method:
          type: string
          description: HTTP metodu
        code:
          type: string
          description: Hata kodu
      required:
        - error

  # ========================================
  # RESPONSES
  # ========================================
  responses:
    BadRequest:
      description: Geçersiz istek parametreleri
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Geçersiz istek parametreleri
            details:
              - field: studentId
                message: UUID formatında olmalıdır
                code: INVALID_UUID
            timestamp: '2025-07-24T14:30:00Z'
    Unauthorized:
      description: Kimlik doğrulama gerekli veya geçersiz
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Authentication required or insufficient permissions
            code: AUTH_REQUIRED
            timestamp: '2025-07-24T14:30:00Z'
...
    ValidationError:
      description: Validasyon hatası
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Invalid request data
            details:
              - field: title
                message: Title is required
                code: REQUIRED_FIELD
              - field: due_date
                message: Invalid due date
                code: INVALID_DATE
            timestamp: '2025-07-24T14:30:00Z'
    InternalError:
      description: Sunucu hatası
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
            code: INTERNAL_ERROR
            timestamp: '2025-07-24T14:30:00Z'
...